# internal/server/routes.gen.go
// Code generated by apigen; DO NOT EDIT.
package server

import (
	"github.com/painscaler/painscaler/internal/analysis"
	"github.com/painscaler/painscaler/internal/logging"
	"github.com/painscaler/painscaler/internal/simulator"
	"github.com/gin-gonic/gin"
	"net/http"
	"strconv"
)

func RegisterRoutes(r *gin.Engine, app *Server) {
	r.POST("/api/v1/refresh", func(c *gin.Context) {
		user := c.GetHeader("Remote-User")
		if err := app.Refresh(user); err != nil {
			writeError(c, err)
			return
		}
		c.Status(http.StatusNoContent)
	})

	r.GET("/api/v1/index", func(c *gin.Context) {
		result, err := app.GetIndex()
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

	r.GET("/api/v1/search", func(c *gin.Context) {
		term := c.Query("q")
		result, err := app.Search(term)
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

	r.GET("/api/v1/segment/:segmentID/policies", func(c *gin.Context) {
		segmentID := c.Param("segmentID")
		result, err := app.PoliciesForSegment(segmentID)
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

	r.GET("/api/v1/reachability", func(c *gin.Context) {
		hostname := c.Query("q")
		result, err := app.WhoCanReach(hostname)
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

	r.GET("/api/v1/reports/orphans", func(c *gin.Context) {
		result, err := app.GetOrphans()
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

	r.GET("/api/v1/reports/overlaps", func(c *gin.Context) {
		result, err := app.GetOverlaps()
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

	r.GET("/api/v1/about", func(c *gin.Context) {
		result := app.GetAbout()
		c.JSON(http.StatusOK, result)
	})

	r.GET("/api/v1/libraries", func(c *gin.Context) {
		result := app.GetLibraries()
		c.JSON(http.StatusOK, result)
	})

	r.POST("/api/v1/simulation/run", func(c *gin.Context) {
		user := c.GetHeader("Remote-User")
		var simCtx simulator.SimContext
		if err := c.ShouldBindJSON(&simCtx); err != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
			return
		}
		result, err := app.RunSimulation(user, simCtx)
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

	r.POST("/api/v1/simulation/compare", func(c *gin.Context) {
		var req CompareRequest
		if err := c.ShouldBindJSON(&req); err != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
			return
		}
		result, err := app.CompareSimulation(req)
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

	r.GET("/api/v1/me", func(c *gin.Context) {
		user := c.GetHeader("Remote-User")
		email := c.GetHeader("Remote-Email")
		groups := c.GetHeader("Remote-Groups")
		name := c.GetHeader("Remote-Name")
		result, err := app.GetMe(user, email, groups, name)
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

	r.POST("/api/v1/telemetry", func(c *gin.Context) {
		user := c.GetHeader("Remote-User")
		var batch logging.TelemetryBatch
		if err := c.ShouldBindJSON(&batch); err != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
			return
		}
		if err := app.PostTelemetry(user, batch); err != nil {
			writeError(c, err)
			return
		}
		c.Status(http.StatusNoContent)
	})

	r.GET("/api/v1/simulation", func(c *gin.Context) {
		limit, err := strconv.ParseInt(c.DefaultQuery("limit", "50"), 10, 64)
		if err != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": "invalid limit"})
			return
		}
		offset, err := strconv.ParseInt(c.DefaultQuery("offset", "0"), 10, 64)
		if err != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": "invalid offset"})
			return
		}
		result, err := app.ListSimulationRuns(limit, offset)
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

	r.GET("/api/v1/simulation/:id", func(c *gin.Context) {
		id, err := strconv.ParseInt(c.Param("id"), 10, 64)
		if err != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": "invalid id"})
			return
		}
		result, err := app.GetSimulationRun(id)
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

	r.DELETE("/api/v1/simulation/:id", func(c *gin.Context) {
		id, err := strconv.ParseInt(c.Param("id"), 10, 64)
		if err != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": "invalid id"})
			return
		}
		if err := app.DeleteSimulationRun(id); err != nil {
			writeError(c, err)
			return
		}
		c.Status(http.StatusNoContent)
	})

	r.GET("/api/v1/simulation/count", func(c *gin.Context) {
		result, err := app.CountSimulationRuns()
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

	r.POST("/api/v1/graph", func(c *gin.Context) {
		var data analysis.GraphQueryBody
		if err := c.ShouldBindJSON(&data); err != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
			return
		}
		result, err := app.GetFlowGraph(data)
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

	r.GET("/api/v1/zpa/client-types", func(c *gin.Context) {
		result, err := app.GetClientTypes()
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

	r.GET("/api/v1/zpa/platforms", func(c *gin.Context) {
		result, err := app.GetPlatforms()
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

	r.GET("/api/v1/zpa/segments", func(c *gin.Context) {
		result, err := app.GetSegments()
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

	r.GET("/api/v1/zpa/segment-groups", func(c *gin.Context) {
		result, err := app.GetSegmentGroups()
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

	r.GET("/api/v1/zpa/app-connectors", func(c *gin.Context) {
		result, err := app.GetAppConnectors()
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

	r.GET("/api/v1/zpa/app-connector-groups", func(c *gin.Context) {
		result, err := app.GetAppConnectorGroups()
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

	r.GET("/api/v1/zpa/access-policies", func(c *gin.Context) {
		result, err := app.GetAccessPolicies()
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

	r.GET("/api/v1/zpa/scim-groups", func(c *gin.Context) {
		result, err := app.GetScimGroups()
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

	r.GET("/api/v1/zpa/scim-attribute-headers", func(c *gin.Context) {
		result, err := app.GetScimAttributeHeaders()
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

	r.GET("/api/v1/zpa/scim-attribute-values", func(c *gin.Context) {
		idpID := c.Query("idpID")
		headerID := c.Query("headerID")
		result, err := app.GetScimAttributeValues(idpID, headerID)
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

	r.GET("/api/v1/zpa/server-groups", func(c *gin.Context) {
		result, err := app.GetServerGroups()
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

	r.GET("/api/v1/zpa/idp-controllers", func(c *gin.Context) {
		result, err := app.GetIdpControllers()
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

	r.GET("/api/v1/zpa/trusted-networks", func(c *gin.Context) {
		result, err := app.GetTrustedNetworks()
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

	r.GET("/api/v1/zpa/posture-profiles", func(c *gin.Context) {
		result, err := app.GetPostureProfiles()
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

	r.GET("/api/v1/zpa/certificates", func(c *gin.Context) {
		result, err := app.GetCertificates()
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

	r.GET("/api/v1/analytics/blast-radius", func(c *gin.Context) {
		targetID := c.Query("targetId")
		targetType := c.Query("targetType")
		result, err := app.GetBlastRadius(targetID, targetType)
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

	r.GET("/api/v1/analytics/policy-shadows", func(c *gin.Context) {
		result, err := app.GetPolicyShadows()
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

	r.GET("/api/v1/analytics/orphan-clusters", func(c *gin.Context) {
		result, err := app.GetOrphanClusters()
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

	r.GET("/api/v1/analytics/domain-overlaps", func(c *gin.Context) {
		result, err := app.GetDomainOverlaps()
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

	r.GET("/api/v1/analytics/connector-load", func(c *gin.Context) {
		result, err := app.GetConnectorLoad()
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

	r.GET("/api/v1/analytics/scim-reach", func(c *gin.Context) {
		result, err := app.GetScimReach()
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

	r.GET("/api/v1/routes", func(c *gin.Context) {
		result, err := app.GetRoutes()
		if err != nil {
			writeError(c, err)
			return
		}
		c.JSON(http.StatusOK, result)
	})

}